Flip order of NSLocation Usage Description checks
Starting in iOS10, showing the user's location on a map requires the NSLocationWhenInUseDescription, even if NSLocationAlwaysInUseDescription is already present in info.plist. There are probably other location features with similar problems. 

It is better to check for the Always In Use description first, because that will enable all location features in an app that contains both descriptions, whereas requesting and enabling on When In Use will only provide a subset. This is especially important because iOS will only present the request dialog the first time an authorization request is made after the app is installed.